+++ test/helpers.js
@@ -1,7 +1,8 @@
-const fs = require('fs')
-const path = require('path')
+const fs = require("fs");
+const path = require("path");
 const targetFolder = path.join(process.cwd(), "static");
-const getPreprocessor = require('../src/index')
+const getPreprocessor = require("../src/index");
+const { defaults } = require("../src/main")
 
 /**
  * Test helper to populate files into the expected directory and will use files
@@ -54,10 +55,13 @@ function cleanFiles() {
  * Convenience function to get directly at the main thing we will be testing
  * @param {*} options Same as the options you'd pass to getPreprocessor
  */
-const getReplaceImages = (options) => (str) => getPreprocessor(options).markup({content:str}).then((obj) => obj.code)
+function getReplaceImages(options) {
+  const preprocessor = getPreprocessor({...defaults, ...options});
+  return str => preprocessor.markup({ content: str }).then(obj => obj.code);
+}
 
 module.exports = {
   cleanFiles,
   populateFiles,
   getReplaceImages
-}
+};